minimally convert vol2img to preprocess_C. Remove _F fft variants
