+++ tests/test_covar2d.py
@@ -12,7 +12,7 @@ from aspire.utils.coor_trans import qrand_rots
 from aspire.utils.filters import RadialCTFFilter
 from aspire.utils.matlab_compat import randn
 from aspire.utils.matrix import anorm
-from aspire.utils.preprocess_F import vol2img as vol2img
+from aspire.utils.preprocess_C import vol2img
 from aspire.volume import Volume
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'saved_test_data')
@@ -52,10 +52,6 @@ class Cov2DTestCase(TestCase):
         # use new methods to generate random rotations and clean images
         sim.rots = qrand_rots(n, seed=0)
         self.imgs_clean = vol2img(vols[0], sim.rots)
-        # RCOPT hack for testing purposes (need to replace vol2img (project?) later)
-        # should probably be using Volume and Image...
-        self.imgs_clean = self.imgs_clean.T
-
 
         self.h_idx = np.array([filters.index(f) for f in sim.filters])
         self.filters = filters
