+++ tests/test_preprocess.py
@@ -5,8 +5,8 @@ import numpy as np
 from scipy.fftpack import fftn, fftshift
 
 from aspire.utils.coor_trans import qrand_rots
-from aspire.utils.preprocess_F import crop_pad, fuzzy_mask, vol2img
-from aspire.utils.preprocess_C import downsample
+from aspire.utils.preprocess_F import crop_pad, fuzzy_mask
+from aspire.utils.preprocess_C import downsample, vol2img
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'saved_test_data')
 
@@ -34,11 +34,10 @@ class PreprocessTestCase(TestCase):
         self.assertTrue(np.allclose(results, vols, atol=1e-7))
 
     def test03Vol2img(self):
-        results = np.load(os.path.join(DATA_DIR, 'clean70SRibosome_down8_imgs32.npy'))
-        vols = np.load(os.path.join(DATA_DIR, 'clean70SRibosome_vol_down8.npy')).T # #RCOPT
-        vols = vols[..., np.newaxis]
+        results = np.load(os.path.join(DATA_DIR, 'clean70SRibosome_down8_imgs32.npy')).T # RCOPT
+        vol = np.load(os.path.join(DATA_DIR, 'clean70SRibosome_vol_down8.npy')).T # RCOPT
         rots = qrand_rots(32, seed=0)
-        imgs_clean = vol2img(vols[..., 0], rots)
+        imgs_clean = vol2img(vol, rots)
         self.assertTrue(np.allclose(results, imgs_clean, atol=1e-7))
 
     def test04FuzzyMask(self):
